♻️ refactor(cli): move type/find command wiring to feature folders
