+++ src/Nupeek.Cli/Features/Find/FindCommandFactory.cs
@@ -1,0 +1,55 @@
+using Nupeek.Core;
+using System.CommandLine;
+using System.CommandLine.Invocation;
+
+namespace Nupeek.Cli;
+
+internal static class FindCommandFactory
+{
+    public static Command Create(GlobalCliOptions globalOptions, Func<PlanRequest, int> runPlan)
+    {
+        var packageOption = new Option<string>("--package", "NuGet package id") { IsRequired = true };
+        packageOption.AddAlias("-p");
+
+        var versionOption = new Option<string?>("--version", "NuGet package version. Defaults to latest.");
+        var tfmOption = new Option<string?>("--tfm", "Target framework moniker. Defaults to auto.");
+        var symbolOption = new Option<string>("--symbol", "Symbol name, e.g. Namespace.Type.Method") { IsRequired = true };
+        var outOption = new Option<string>("--out", "Output directory (e.g. deps-src)") { IsRequired = true };
+        var formatOption = new Option<string>("--format", () => "text", "Output format: text (default) or json.");
+        var emitOption = new Option<string>("--emit", () => "files", "Emit mode: files (default) or agent.");
+        var maxCharsOption = new Option<int>("--max-chars", () => 12000, "Max inline source chars for --emit agent.");
+
+        var command = new Command("find", "Resolve symbol to type and decompile that type.");
+        command.AddOption(packageOption);
+        command.AddOption(versionOption);
+        command.AddOption(tfmOption);
+        command.AddOption(symbolOption);
+        command.AddOption(outOption);
+        command.AddOption(formatOption);
+        command.AddOption(emitOption);
+        command.AddOption(maxCharsOption);
+
+        command.SetHandler((InvocationContext context) =>
+        {
+            var parse = context.ParseResult;
+            var symbol = parse.GetValueForOption(symbolOption)!;
+            Environment.ExitCode = runPlan(new PlanRequest(
+                Command: "find",
+                Package: parse.GetValueForOption(packageOption)!,
+                Version: parse.GetValueForOption(versionOption) ?? "latest",
+                Tfm: parse.GetValueForOption(tfmOption) ?? "auto",
+                Type: SymbolParser.ToTypeName(symbol),
+                OutDir: parse.GetValueForOption(outOption)!,
+                Verbose: parse.GetValueForOption(globalOptions.Verbose),
+                Quiet: parse.GetValueForOption(globalOptions.Quiet),
+                DryRun: parse.GetValueForOption(globalOptions.DryRun),
+                Format: parse.GetValueForOption(formatOption) ?? "text",
+                Emit: parse.GetValueForOption(emitOption) ?? "files",
+                MaxChars: parse.GetValueForOption(maxCharsOption),
+                Progress: parse.GetValueForOption(globalOptions.Progress) ?? "auto",
+                SourceSymbol: symbol));
+        });
+
+        return command;
+    }
+}
